+++ q2_matchmaker/tests/test_stan.py
@@ -1,6 +1,6 @@
 import unittest
 import numpy as np
-from q2_differential._stan import (
+from q2_matchmaker._stan import (
     _case_control_sim, _case_control_full,
     _case_control_data, _case_control_single,
 )
